refactor: placement of cmps, hooks & other small stuff updated
