+++ apps/user-client/src/main.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
-import App from './App';
+import { App } from './App';
 import { api } from './redux-toolkit/api';
 import { store } from './redux-toolkit/store';
