+++ apps/user-client/src/pages/update-card/UpdateCard.tsx
@@ -1,7 +1,3 @@
 export const UpdateCard = () => {
-  return (
-    <>
-      <h1>This is a Update card page</h1>
-    </>
-  );
+  return <h1>This is a Update card page</h1>;
 };
