+++ libs/shared/ui/src/assets/mocks/avatar1.jpg
