+++ libs/shared/ui/src/assets/mocks/avatar2.jpg
