+++ libs/shared/ui/src/assets/mocks/avatar3.jpg
