+++ libs/shared/ui/src/assets/mocks/avatar4.jpg
