+++ libs/shared/ui/src/assets/mocks/avatar5.jpg
