+++ libs/shared/ui/src/assets/mocks/avatar6.jpg
