+++ libs/shared/ui/src/assets/mocks/avatar7.jpg
