+++ libs/shared/ui/src/assets/mocks/avatar8.jpg
