+++ libs/shared/ui/src/assets/mocks/avatarImages.ts
