+++ libs/shared/ui/src/assets/mocks/index.ts
