+++ libs/shared/ui/src/components/avatar/index.ts
@@ -1,0 +1,1 @@
+export * from './ChangeAvatar';
