+++ libs/shared/ui/src/components/button/Button.tsx
@@ -2,6 +2,4 @@ import { ButtonBaseStyles } from './Button.styled';
 
 export type ButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement>;
 
-const Button = (props: ButtonProps) => <ButtonBaseStyles {...props} />;
-
-export default Button;
+export const Button = (props: ButtonProps) => <ButtonBaseStyles {...props} />;
