+++ libs/shared/ui/src/components/form/index.ts
@@ -1,0 +1,2 @@
+export * from './Form';
+export * from './hooks';
