+++ libs/shared/ui/src/components/index.ts
@@ -1,6 +1,11 @@
+export * from './avatar';
 export * from './button';
 export * from './card';
-export * from './navigation';
 export * from './error';
+export * from './form';
 export * from './image';
+export * from './input';
 export * from './loader';
+export * from './navigation';
+export * from './text-area';
+export * from './validation-error';
