+++ libs/shared/ui/src/components/text-area/TextArea.styled.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { space } from '../../../variables';
+import { space } from '../../variables';
 import { StyledFormControl } from '../input/Input.styled';
 
 export const StyledTextAreaFormControl = styled(StyledFormControl)<{ hasLabel: boolean }>`
