+++ libs/shared/ui/src/components/text-area/index.ts
@@ -1,0 +1,1 @@
+export * from './TextArea';
