+++ libs/shared/ui/src/utils/error-casting.tsx
